requirements.txt
fastapi==0.127.0
uvicorn==0.24.0
ezdxf==1.4.2
pydantic==2.12.5
python-multipart==0.0.6
pytest==8.4.2
pytest-asyncio==0.23.0
httpx==0.28.1
fastapi-cache2==0.2.2
python-dotenv==1.0.0
